removed connectors_test to datasets_test, added RIME and CASA conj to conventions.
