--- docs/api.md.backup
@@ -1,85 +1,0 @@
-(api-reference-label)=
-
-# API
-
-This page documents all of the available components of the MPoL package. If you do not see something that you think should be documented, please raise an [issue](https://github.com/iancze/MPoL/issues).
-
-## Utilities
-
-```{eval-rst}
-.. automodule:: mpol.utils
-
-```
-
-## Coordinates
-
-```{eval-rst}
-.. automodule:: mpol.coordinates
-```
-
-## Geometry
-
-```{eval-rst}
-.. automodule:: mpol.geometry
-```
-
-## Gridding
-
-```{eval-rst}
-.. automodule:: mpol.gridding
-```
-
-## Datasets
-
-```{eval-rst}
-.. automodule:: mpol.datasets
-```
-
-## Images
-
-```{eval-rst}
-.. automodule:: mpol.images
-```
-
-## Fourier
-
-```{eval-rst}
-.. automodule:: mpol.fourier
-
-```
-
-## Precomposed Modules
-
-For convenience, we provide some "precomposed" [modules](https://pytorch.org/docs/stable/notes/modules.html) which may be useful for simple imaging or modeling applications. In general, though, we encourage you to compose your own set of layers if your application requires it. The source code for a precomposed network can provide useful a starting point. We also recommend checking out the PyTorch documentation on [modules](https://pytorch.org/docs/stable/notes/modules.html).
-
-```{eval-rst}
-.. automodule:: mpol.precomposed
-```
-
-
-## Training and testing
-
-```{eval-rst}
-.. automodule:: mpol.training
-
-```
-
-## Cross-validation
-
-```{eval-rst}
-.. automodule:: mpol.crossval
-
-```
-
-## Analysis
-
-```{eval-rst}
-.. automodule:: mpol.onedim
-
-```
-
-## Plotting
-
-```{eval-rst}
-.. automodule:: mpol.plot
-```
